Change ints to long longs (#47)

+++ src/cpp_src/pb_methods_and_measures/Greedy.h
@@ -7,7 +7,8 @@
 
 std::vector<ProjectEmbedding> greedy(const Election &election, const ProjectComparator &tie_breaking);
 
-std::optional<int> cost_reduction_for_greedy(const Election &election, int p, const ProjectComparator &tie_breaking);
+std::optional<long long> cost_reduction_for_greedy(const Election &election, int p,
+                                                   const ProjectComparator &tie_breaking);
 
 std::optional<int> optimist_add_for_greedy(const Election &election, int p, const ProjectComparator &tie_breaking);
